Fix: Total3 Route fixed
